+++ fiddles/angular2-cli/fiddle-0017-SlickGridTreePlus/src/app/components/index.ts
@@ -1,2 +1,0 @@
-export * from './tree-grid/index';
-export * from './shared/index';