+++ fiddles/angular2-cli/fiddle-0017-SlickGridTreePlus/src/app/components/tree-grid/enums/index.ts
@@ -1,2 +1,0 @@
-export * from './tree-grid-events.enum';
-export * from './tree-grid-cmds.enum';