+++ fiddles/angular2-cli/fiddle-0017-SlickGridTreePlus/src/app/components/tree-grid/interfaces/index.ts
@@ -1,4 +1,0 @@
-export * from './tree-grid-node.interface';
-export * from './tree-grid.interface';
-export * from './tree-grid-event.interface';
-