+++ fiddles/angular2-cli/fiddle-0017-SlickGridTreePlus/src/app/components/tree-grid/interfaces/tree-grid-event.interface.ts
@@ -1,8 +1,0 @@
-import {
-  TreeGridEvents
-} from '../enums/index';
-
-export interface TreeGridEventInterface {
-  type: TreeGridEvents;
-  data: any;
-}