+++ fiddles/angular2-cli/fiddle-0017-SlickGridTreePlus/src/app/components/tree-grid/services/index.ts
@@ -1,3 +1,0 @@
-export * from './tree-grid-events.service';
-export * from './tree-grid-resize.service';
-export * from './tree-grid-node.service';