+++ fiddles/angular2-cli/fiddle-0017-SlickGridTreePlus/src/app/shared/index.ts
@@ -1,3 +1,0 @@
-export * from './data.service';
-export * from './state.service';
-